fix: member findById Integer로 변경

+++ Backend/src/main/java/com/whitemind/cogit/member/repository/MemberRepository.java
@@ -3,6 +3,6 @@ package com.whitemind.cogit.member.repository;
 import com.whitemind.cogit.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Member findMembersByMemberId(Long memberId);
+public interface MemberRepository extends JpaRepository<Member, Integer> {
+    Member findMembersByMemberId(int memberId);
 }
